Changelog 2/18/2023
The code has undergone several changes: the project structure, file and directory names have been modified. It's still suffering, as i have no idea how to better name the directories, classes, interfaces, etc.
The main changes are:
IDetectedAccount was removed.
An extension was added for ISteamID.
Unnecessary code was moved from the ISteamID interface in SteamIDConverter.
The FindAccountsAsync methods now return IEnumerable<ISteamID>.
Appmanifest and Appworkshop scan-methods were separated instead of having a single method, which increased the speed of these methods on my processor from 600 ms (one method) to 10 ms (20 ms in total).
Console output is now prettier.
Exception throw was removed in case the SteamClient was not found - after all, it's possible to find an account in the registry by chance.